# docs/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact: {}
  description: REST Api of the CollabTED project.
  title: CollabTED
  version: "1.0"
paths:
...
  /oauth/facebook/callback:
    get:
      consumes:
      - application/json
      parameters:
      - description: The OAuth2 authorization code
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Handles Facebook OAuth2 callback
      tags:
      - auth
...
  /oauth/google/callback:
    get:
      consumes:
      - application/json
      parameters:
      - description: The OAuth2 authorization code
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Handles Google OAuth2 callback
      tags:
      - auth
...
swagger: "2.0"